MLHR-1758 #resolve moved RubyOperator to contrib because of transitive dependency that can break stram webservice
